requirements.txt
absl-py==2.1.0
accelerate==0.30.1
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blobfile==2.1.1
bottleneck==1.3.7
cffi==1.16.0
charset-normalizer==2.0.4
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.27.2
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.5.0
gdown==5.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.1
importlib-metadata==7.1.0
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
jedi==0.19.1
jinja2==3.1.3
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.1
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.0
jupyter-server-terminals==0.5.3
jupyterlab==4.2.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.1
kiwisolver==1.4.4
lmdb==1.4.1
lxml==4.9.4
markupsafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
ml-collections==0.1.1
nest-asyncio==1.6.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
notebook==7.2.0
notebook-shim==0.2.4
numexpr==2.8.7
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.9.0.80
overrides==7.7.0
packaging==24.0
pandas==2.1.1
pillow==10.3.0
ply==3.11
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pygments==2.18.0
pyparsing==3.0.9
python-dateutil==2.9.0post0
python-json-logger==2.0.7
pytz==2024.1
pyyaml==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.0
send2trash==1.8.3
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
terminado==0.18.1
tifffile==2024.5.10
timm==1.0.3
tinycss2==1.3.0
tokenizers==0.19.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
unicodedata2==15.1.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
zipp==3.18.2
